pyproject.toml
[tool.poetry]
name = "omnizart"
version = "0.1.1"
description = "Omniscient Mozart, being able to transcribe everything in the music."
readme = "README.md"
authors = ["BreezeWhite <freedombluewater@gmail.com>"]
license = "MIT"
repository = "https://github.com/Music-and-Culture-Technology-Lab/omnizart"
documentation = "https://music-and-culture-technology-lab.github.io/omnizart-doc/"
homepage = "https://sites.google.com/view/mctl/home"
keywords = ["automatic music transcription", "drum transcription", "vocal transcription", "beat tracking"]
classifiers = [
    "Topic :: Multimedia :: Sound/Audio :: MIDI",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Version Control :: Git",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",

]
exclude = ["weights*"]

[tool.poetry.dependencies]
python = ">=3.6.1,<3.9"
click = "^7.1.2"
pretty_midi = "^0.2.9"
librosa = "^0.8.0"
tensorflow = "^2.3.0"
madmom = "^0.16.1"
opencv-python = "^4.4.0"
pyyaml = "^5.3.1"
jsonschema = "^3.2.0"
numba = "0.48"
tqdm = "^4.49.0"
pyfluidsynth = "^1.2.5"
vamp = "^1.1.0"
urllib3 = "^1.25.11"
spleeter = "^2.0.1"
mir_eval = "^0.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
pylint = "^2.5.3"
black = "^19.10b0"
numpydoc = "^1.1.0"
yapf = "^0.30.0"
sphinx-press-theme = "^0.5.1"
pytest-cov = "^2.10.1"
pytest-mock = "^3.3.1"
sphinx-click = "^2.5.0"

[tool.poetry.scripts]
omnizart = 'omnizart.cli.cli:entry'

[tool.black]
line-length = 120
target-version = ['py36']

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

